requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
passlib[bcrypt]==1.7.4
opencv-contrib-python==4.10.0.84
numpy>=2.0.0
Pillow==10.4.0
python-dotenv==1.0.1

